Added function to check admission and discharge dates
Added full checkDates function, as part of the Data management and data checks. It takes the (minimal) data set as input, and depending on the options outputs the corrected database.

It doesn't correct anything using the default options, instead gives an error with suggested options for the next try.

Up to now, I always worked with POSIXct as date type. It uses the lubridate library to convert the dates. I'm open to using other date types if there's a better way