Adjusted edgelist_from_patient_database to fit workflow
I included the admission and discharge dates in a new condition (C4), depending on the threshold window for it to be called a transfer.

I commented C2 & C3 out, as these should be conditional on the variable being available, and the choice whether the function should count only direct transmssions.