Introduced deleteErrorRecords function and followed through variable naming
Because deletion of records or patients (whatever specified by the user) is done a number of times in the script, I placed it in its own function, and just called the function when deletion was needed.

This also meant I had to follow through with the naming convention of the variables. Each function therefore takes the names of the required variable, such that throughout the script, the database retains its original variable naming.

Extra minor edit: the script stopped when parsing of the dates failed. I now give the option to delete the records/patients and continue the rest of the script. Also, date conversion didn't seem to work.

*I know it's a lot of edits for one commit, going a bit against the 'best practice' for git, but I went down this rabbit hole this morning after trying the script out on a dirty dataset (deleted 15k records out of 99k total). In trying to solve it, I needed to rework quite a bit of the data management side of it.