Streamlining HospiNet to work with igraph not matrix
I worked most of the references to the full matrix out of the HospiNet class (object?). This because the full matrix might get quite large with increasing number of hospitals.

Put a message in for when the matrix is created, but most metrics are extracted from the weighted igraph. Give same results with my tested datasets.