Big workflow update, moving to HospiNet class
I included all preferred metrics as variables of the hospiNet class. The variable/function $metricsTable will calculate all per hospital metrics and return it as a data.table. Furthermore, statistics on the database are/can also stored in the HospiNet object

For me, the advantage of doing it this way, is that we can save (and share) the HospiNet object in .rds format, without having to report on each metric individually.

HospiNet: see above. Also: initialisation step includes the option to input summary statistics of the database.

NetworkBuilding: included summary statistics of database and hospitals in the hospinet_from_patient_database() function, which passes it on to the HospiNet object. This way, statistics on numbers of admissions/patients/etc per hospital end up in the same table as the network metrics for each hospital. As network centrality and number of admitted patients are related, best to present them in the same place.

NetworkAnalysis: Primarily reworked to treat cluster (community detecting) as any other metric.
Also small bugfix: used local variable "myNet" where I should have used function input variable "graph" (Thanks to Xanthi Andrianou for noticing)

DataSummary: For consistency, changed column name from hID to node.