New plot type: circular_network
Also known as spaghetti plot, or edge bundling plot. It's a plot I've used on a number of occasions. Making it generalisable for any network is a bit of a challenge, but seems to give reasonable plots at the moment.

Still need to add colouring to it, but flow within and between clusters is already quite nicely visible.