Fixing bug in HospiNet when window_threshold = 0
The bug was with the get_graphs_bycluster() and get_hubs_bycluster()
functions. In get_graphs_bycluster(), the clusters with only one
member were removed. So if there were only clusters with one member,
the function returned an empty list, which caused the error in
get_hubs_bycluster(). Now the graphs are computed for all clusters,
but a warning is raised if the cluster has only one member.