Different handling of "type III" overlap
Type III overlaps are overlaps with the same discharge dates in both records. Until now this resulted in a single day admission at the discharge day, but I think it's best to exclude these, and treat them as "Type I" overlaps (i.e. partial overlap, only needing adjustment of the discharge date of the forst record).