Remove stats on dataset from HospiNet
I've removed the summary statistics on the data set from the HospiNet object, and inserted them into the report returned by checkBase().

In this way, any information on the entire dataset is returned by checkbase(), and any information on the facilities (nodes in the network) is kept in the HospiNet object.