update beta branch with current master (#18)
* Fixing bug in computation of subjects and facilities in DataSummary

and now computing TBA in days

* Explicitly write option that includes window_threshold in the cutoff

incbounds = TRUE was the default anyway, but makes it explicit when
reading the code

* Different handling of "type III" overlap

Type III overlaps are overlaps with the same discharge dates in both records. Until now this resulted in a single day admission at the discharge day, but I think it's best to exclude these, and treat them as "Type I" overlaps (i.e. partial overlap, only needing adjustment of the discharge date of the forst record).

* Remove stats on dataset from HospiNet

I've removed the summary statistics on the data set from the HospiNet object, and inserted them into the report returned by checkBase().

In this way, any information on the entire dataset is returned by checkbase(), and any information on the facilities (nodes in the network) is kept in the HospiNet object.

* fix documentation pID -> sID, hID -> fID

* - adjust checkBase documentation with an example
- update website
- dev version 0.9.1.1

* bugfix: facility degrees were left out of metricstable

Inlcuded the degrees again. Probably mistake during previous commit, removing database stats from hospinet

* add code coverage capabilities

* display coverage and build status on Readme.md

* - add tests for HospiNet plotting functions
- add appveyor (to test windows build in addition to travis)
- global spell check (add word list for specific terms)
- remove all remaining Notes at build (use globalvariables in zzz.R, add .Rbuildignore for non-standard package files)

* - update website

* - add "ggraph" to Description Imports

* - cover plot of hist degree in test
- fix documentation for HospiNet (some fields are not implemented yet

* First running version of the shiny app

* fixed .RD

* - fix warnings and dependencies
- update website and increase dev version number

* update version
small fix for appveyor
caching for travis

* -test

* - another try at fixing appveyor CI

* Deduplication of records only on key variables.

Follwoing Michael's suggestion, we can deduplicate without auxiliary information. The algorithm correcting the overlapping admissions is unable to take aux data into account anyway, and this improves the speed of the algorithm.

* - fix typo
- add a goto "Construct Net" button

* - fix: shortened paths for svg figures

* - add progress bar to network building

* - add circular network

* - add notification system to the shiny interface to keep track of what has been done in the previous tabs

* - add the parameter as input for the circular plot
- make "remove loops" True by default -> possible bug to investigate

* - update test plot image after simplification of circular plot function

* - add fake data interface
- slighlty modify spaghetti plot to make it work with poorly connected graph

* - fix bug in "get_metrics"
- add test for "get_metrics"

* - add shiny tests

* - add shiny tests

* -update travis for shinytest

* - add phantomjs to appveyor

* fix appveyor

* - remove packrate from travis