Added probability-based network reconstruction
This method is based on the original Donker et al. PLoS CB paper (2010), which takes all previous admissions into account. It weights each movement by the probability of transmission given the LOS in hospital A, time between discharge from A and admission to B, and the LOS in hospital B

It is a parametric approach, using three parameters:
- the rate of transmission to the focal patient in hospital A
- the rate of loss of colonisation
- the rate of transmission from the focal patient to any of the patients in hospital B

-I needed to change the construction of el_aggr from el_long for the other methods as well, because el_long is defined differently for the new method (it's a probability, not a single count for a single movement).

I have checked if this works, and don't seem to run into problems, but there is a chance I missed something... Keep in mind for future bugs involving el_long or el_aggr.